# src/docs/swagger.yaml
definitions:
...
  github_com_nazarow_golang-clean_api_dto.CarModelColorResponse:
    properties:
      color:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.ColorResponse"
      id:
        type: integer
    type: object
...
  github_com_nazarow_golang-clean_api_dto.CarModelImageResponse:
    properties:
      carModelId:
        type: integer
      id:
        type: integer
      image:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.FileResponse"
      isMainImage:
        type: boolean
    type: object
...
  github_com_nazarow_golang-clean_api_dto.ColorResponse:
    properties:
      hexCode:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.CreateCarModelColorRequest:
    properties:
      carModelId:
        type: integer
      colorId:
        type: integer
    required:
      - carModelId
      - colorId
    type: object
...
  github_com_nazarow_golang-clean_api_dto.CreateCarModelImageRequest:
    properties:
      carModelId:
        type: integer
      imageId:
        type: integer
      isMainImage:
        type: boolean
    required:
      - carModelId
      - imageId
    type: object
...
  github_com_nazarow_golang-clean_api_dto.FileResponse:
    properties:
      description:
        type: string
      directory:
        type: string
      id:
        type: integer
      mimeType:
        type: string
      name:
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description:
          contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_CarModelColorResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelColorResponse"
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_CarModelImageResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelImageResponse"
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.Filter"
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.Sort"
        type: array
    type: object
...
  github_com_nazarow_golang-clean_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
  github_com_nazarow_golang-clean_api_dto.UpdateCarModelColorRequest:
    properties:
      carModelId:
        type: integer
      colorId:
        type: integer
    type: object
...
  github_com_nazarow_golang-clean_api_dto.UpdateCarModelImageRequest:
    properties:
      isMainImage:
        type: boolean
    type: object
...
  github_com_nazarow_golang-clean_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.ResultCode"
      success:
        type: boolean
      validationErrors:
        items:
          $ref: "#/definitions/github_com_nazarow_golang-clean_api_validations.ValidationError"
        type: array
    type: object
  github_com_nazarow_golang-clean_api_helper.ResultCode:
    enum:
      - 0
      - 40001
      - 40101
      - 40301
      - 40401
      - 42901
      - 42902
      - 50001
      - 50002
    type: integer
    x-enum-varnames:
      - Success
      - ValidationError
      - AuthError
      - ForbiddenError
      - NotFoundError
      - LimiterError
      - OtpLimiterError
      - CustomRecovery
      - InternalError
  github_com_nazarow_golang-clean_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/car-model-colors/:
    post:
      consumes:
        - application/json
      description: Create a CarModelColor
      parameters:
        - description: Create a CarModelColor
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CreateCarModelColorRequest"
      responses:
        "201":
          description: CarModelColor response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Create a CarModelColor
      tags:
        - CarModelColors
  /v1/car-model-colors/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a CarModelColor
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Delete a CarModelColor
      tags:
        - CarModelColors
    get:
      consumes:
        - application/json
      description: Get a CarModelColor
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get a CarModelColor
      tags:
        - CarModelColors
    put:
      consumes:
        - application/json
      description: Update a CarModelColor
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
        - description: Update a CarModelColor
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.UpdateCarModelColorRequest"
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Update a CarModelColor
      tags:
        - CarModelColors
  /v1/car-model-colors/get-by-filter:
    post:
      consumes:
        - application/json
      description: Get CarModelColors
      parameters:
        - description: Request
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter"
      responses:
        "200":
          description: CarModelColor response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_CarModelColorResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get CarModelColors
      tags:
        - CarModelColors
...
  /v1/car-model-images/:
    post:
      consumes:
        - application/json
      description: Create a CarModelImage
      parameters:
        - description: Create a CarModelImage
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CreateCarModelImageRequest"
      responses:
        "201":
          description: CarModelImage response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelImageResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Create a CarModelImage
      tags:
        - CarModelImages
  /v1/car-model-images/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a CarModelImage
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Delete a CarModelImage
      tags:
        - CarModelImages
    get:
      consumes:
        - application/json
      description: Get a CarModelImage
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
      responses:
        "200":
          description: CarModelImage response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelImageResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get a CarModelImage
      tags:
        - CarModelImages
    put:
      consumes:
        - application/json
      description: Update a CarModelImage
      parameters:
        - description: Id
          in: path
          name: id
          required: true
          type: integer
        - description: Update a CarModelImage
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.UpdateCarModelImageRequest"
      responses:
        "200":
          description: CarModelImage response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.CarModelImageResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Update a CarModelImage
      tags:
        - CarModelImages
  /v1/car-model-images/get-by-filter:
    post:
      consumes:
        - application/json
      description: Get CarModelImages
      parameters:
        - description: Request
          in: body
          name: Request
          required: true
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PaginationInputWithFilter"
      responses:
        "200":
          description: CarModelImage response
          schema:
            allOf:
              - $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
              - properties:
                  result:
                    $ref: "#/definitions/github_com_nazarow_golang-clean_api_dto.PagedList-github_com_nazarow_golang-clean_api_dto_CarModelImageResponse"
                type: object
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/github_com_nazarow_golang-clean_api_helper.BaseHttpResponse"
      security:
        - AuthBearer: []
      summary: Get CarModelImages
      tags:
        - CarModelImages
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"